Move Theme to the components package
