+++ app/src/main/java/eu/epitech/reyditech/components/Theme.kt
@@ -1,4 +1,4 @@
-package eu.epitech.reyditech
+package eu.epitech.reyditech.components
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
